Move exceptions to top-level
