+++ optimade/filtertransformers/base_transformer.py
@@ -11,7 +11,7 @@ from typing import Any, Dict, Optional, Type
 
 from lark import Transformer, Tree, v_args
 
-from optimade.server.exceptions import BadRequest
+from optimade.exceptions import BadRequest
 from optimade.server.mappers import BaseResourceMapper
 from optimade.warnings import UnknownProviderProperty
