+++ optimade/filtertransformers/mongo.py
@@ -11,8 +11,8 @@ from typing import Any, Dict, List, Union
 
 from lark import Token, v_args
 
+from optimade.exceptions import BadRequest
 from optimade.filtertransformers.base_transformer import BaseTransformer, Quantity
-from optimade.server.exceptions import BadRequest
 from optimade.warnings import TimestampNotRFCCompliant
 
 __all__ = ("MongoTransformer",)
