+++ optimade/server/entry_collections/entry_collections.py
@@ -5,10 +5,10 @@ from typing import Any, Dict, Iterable, List, Set, Tuple, Type, Union
 
 from lark import Transformer
 
+from optimade.exceptions import BadRequest, Forbidden, NotFound
 from optimade.filterparser import LarkParser
 from optimade.models.entries import EntryResource
 from optimade.server.config import CONFIG, SupportedBackend
-from optimade.server.exceptions import BadRequest, Forbidden, NotFound
 from optimade.server.mappers import BaseResourceMapper
 from optimade.server.query_params import EntryListingQueryParams, SingleEntryQueryParams
 from optimade.warnings import (
